Fix: "Import" steps warns when users haven't selected/uploaded files
Previously, users could choose a file from their system but forget to upload it,
and the CWA would be fine. Now the CWA warns when there are files left in the
upload queue. I also added a warning when both the upload queue and list of
uploaded files are empty.

Closes #41